@@ -22,58 +22,57 @@
         "migrate": "typeorm migration:run"
     },
     "dependencies": {
-        "@nestjs/axios": "^0.0.5",
-        "@nestjs/common": "^8.2.6",
-        "@nestjs/config": "^1.1.7",
-        "@nestjs/core": "^8.2.6",
-        "@nestjs/jwt": "^8.0.0",
-        "@nestjs/passport": "^8.1.0",
-        "@nestjs/platform-express": "^8.2.6",
-        "@nestjs/schedule": "^1.0.2",
-        "@nestjs/swagger": "^5.2.0",
-        "@nestjs/typeorm": "^8.0.3",
-        "blizzapi": "^2.1.1",
+        "@nestjs/axios": "^0.1.0",
+        "@nestjs/common": "^9.0.6",
+        "@nestjs/config": "^2.2.0",
+        "@nestjs/core": "^9.0.6",
+        "@nestjs/jwt": "^9.0.0",
+        "@nestjs/passport": "^9.0.0",
+        "@nestjs/platform-express": "^9.0.6",
+        "@nestjs/schedule": "^2.1.0",
+        "@nestjs/swagger": "^6.0.4",
+        "@nestjs/typeorm": "^9.0.0",
+        "blizzapi": "^2.3.0",
         "class-transformer": "^0.5.1",
         "class-validator": "^0.13.2",
         "cookie-parser": "^1.4.6",
-        "passport": "^0.5.2",
+        "passport": "^0.6.0",
         "passport-jwt": "^4.0.0",
         "passport-oauth2": "^1.6.1",
         "reflect-metadata": "^0.1.13",
         "rimraf": "^3.0.2",
-        "rxjs": "^7.5.4",
-        "sqlite3": "^5.0.2",
-        "swagger-ui-express": "^4.3.0",
-        "typeorm": "^0.2.34",
+        "rxjs": "^7.5.6",
+        "sqlite3": "^5.0.10",
+        "typeorm": "^0.3.7",
         "uuid": "^8.3.2"
     },
     "devDependencies": {
-        "@nestjs/cli": "^8.2.0",
-        "@nestjs/schematics": "^8.0.6",
-        "@nestjs/testing": "^8.3.1",
-        "@types/cookie-parser": "^1.4.2",
-        "@types/cron": "^1.7.3",
+        "@nestjs/cli": "^9.0.0",
+        "@nestjs/schematics": "^9.0.1",
+        "@nestjs/testing": "^9.0.6",
+        "@types/cookie-parser": "^1.4.3",
+        "@types/cron": "^2.0.0",
         "@types/express": "^4.17.13",
-        "@types/jest": "^27.4.0",
-        "@types/node": "^17.0.17",
+        "@types/jest": "^28.1.6",
+        "@types/node": "^18.6.2",
         "@types/passport-jwt": "^3.0.6",
         "@types/passport-oauth2": "^1.4.11",
-        "@types/supertest": "^2.0.11",
+        "@types/supertest": "^2.0.12",
         "@types/uuid": "^8.3.4",
-        "@typescript-eslint/eslint-plugin": "^5.11.0",
-        "@typescript-eslint/parser": "^5.11.0",
-        "eslint": "^8.9.0",
-        "eslint-config-prettier": "^8.3.0",
-        "eslint-plugin-prettier": "^4.0.0",
+        "@typescript-eslint/eslint-plugin": "^5.31.0",
+        "@typescript-eslint/parser": "^5.31.0",
+        "eslint": "^8.20.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-prettier": "^4.2.1",
         "eslint-plugin-require": "^0.0.1",
-        "jest": "^27.5.1",
-        "prettier": "^2.5.1",
-        "supertest": "^6.2.2",
-        "ts-jest": "^27.1.3",
-        "ts-loader": "^9.2.6",
-        "ts-node": "^10.5.0",
-        "tsconfig-paths": "^3.12.0",
-        "typescript": "^4.5.5"
+        "jest": "^28.1.3",
+        "prettier": "^2.7.1",
+        "supertest": "^6.2.4",
+        "ts-jest": "^28.0.7",
+        "ts-loader": "^9.3.1",
+        "ts-node": "^10.9.1",
+        "tsconfig-paths": "^4.0.0",
+        "typescript": "^4.7.4"
     },
     "jest": {
         "moduleFileExtensions": [
