@@ -3,29 +3,28 @@
     "version": "0.2.5",
     "private": true,
     "dependencies": {
-        "@emotion/react": "^11.8.2",
-        "@emotion/styled": "^11.8.1",
-        "@material-ui/core": "^5.0.0-beta.5",
-        "@material-ui/data-grid": "^4.0.0-alpha.37",
-        "@material-ui/icons": "^5.0.0-beta.5",
-        "@material-ui/styles": "^5.0.0-beta.5",
-        "@testing-library/jest-dom": "^5.16.2",
+        "@emotion/react": "^11.9.3",
+        "@emotion/styled": "^11.9.3",
+        "@mui/icons-material": "^5.8.4",
+        "@mui/material": "^5.9.2",
+        "@mui/styles": "^5.9.2",
+        "@mui/x-data-grid": "^5.15.0",
+        "@testing-library/jest-dom": "^5.16.4",
         "@testing-library/react": "^12.1.4",
         "@testing-library/user-event": "^13.5.0",
-        "@types/jest": "^27.4.1",
-        "@types/node": "^17.0.21",
-        "@types/react": "^17.0.40",
-        "@types/react-dom": "^17.0.13",
+        "@types/jest": "^28.1.6",
+        "@types/node": "^18.6.2",
+        "@types/react": "^18.0.15",
+        "@types/react-dom": "^18.0.6",
         "@types/react-router-dom": "^5.3.3",
-        "chart.js": "^3.5.1",
-        "lorem-ipsum": "^2.0.4",
-        "react": "^17.0.2",
-        "react-chartjs-2": "^3.0.4",
-        "react-dom": "^17.0.2",
+        "chart.js": "^3.8.2",
+        "react": "^18.2.0",
+        "react-chartjs-2": "^4.3.1",
+        "react-dom": "^18.2.0",
         "react-helmet": "^6.1.0",
-        "react-router-dom": "^6.2.2",
-        "react-scripts": "^5.0.0",
-        "typescript": "^4.6.2",
+        "react-router-dom": "^6.3.0",
+        "react-scripts": "^5.0.1",
+        "typescript": "^4.7.4",
         "web-vitals": "^2.1.4"
     },
     "scripts": {
@@ -60,10 +59,10 @@
     },
     "devDependencies": {
         "@types/react-helmet": "^6.1.5",
-        "@typescript-eslint/eslint-plugin": "^5.14.0",
-        "eslint": "^8.10.0",
+        "@typescript-eslint/eslint-plugin": "^5.31.0",
+        "eslint": "^8.20.0",
         "eslint-config-prettier": "^8.5.0",
-        "eslint-plugin-prettier": "^4.0.0",
-        "prettier": "^2.5.1"
+        "eslint-plugin-prettier": "^4.2.1",
+        "prettier": "^2.7.1"
     }
 }
