@@ -12,8 +12,8 @@ import {
     ListItemText,
     TextField,
     Typography,
-} from "@material-ui/core";
-import { Delete, ManageAccounts, AddCircle, ArrowDownward, ArrowUpward } from "@material-ui/icons";
+} from "@mui/material";
+import { Delete, ManageAccounts, AddCircle, ArrowDownward, ArrowUpward } from "@mui/icons-material";
 import React from "react";
 import {
     Collaborator,
