@@ -1,12 +1,17 @@
-import { Column, Entity, JoinColumn, OneToOne, UpdateDateColumn } from "typeorm";
+import { Column, Entity, JoinColumn, OneToOne, PrimaryColumn, UpdateDateColumn } from "typeorm";
 import { Raider } from "./raider.entity";
 
 @Entity()
 export class CachedOverview {
+    @PrimaryColumn({
+        type: "uuid",
+        length: 32,
+    })
+    raiderId: string;
+
     @OneToOne(() => Raider, {
         onDelete: "CASCADE",
         onUpdate: "CASCADE",
-        primary: true,
     })
     @JoinColumn({ name: "raiderId" })
     raider: Raider;
