@@ -22,7 +22,7 @@ export class CollaboratorsService {
 
         return this.collaboratorsRepository.find({
             where: {
-                raidTeam: raidTeam,
+                raidTeam: { id: raidTeam.id },
             },
         });
     }
@@ -45,9 +45,11 @@ export class CollaboratorsService {
             role: role,
         });
 
-        return this.collaboratorsRepository.findOne({
+        return this.collaboratorsRepository.findOneBy({
             battletag: battletag.toLowerCase(),
-            raidTeam: raidTeam,
+            raidTeam: {
+                id: raidTeam.id,
+            },
         });
     }
 
