@@ -22,7 +22,7 @@ import { RaidTierConfiguration } from "src/commons/raid-tier-configuration";
 import { CachedOverview } from "src/entities/cached-overview.entity";
 import { Cron, CronExpression } from "@nestjs/schedule";
 import { User } from "src/entities/user.entity";
-import { RegionName } from "blizzapi";
+import { RegionNameEnum } from "blizzapi";
 import { AccessService } from "src/raid-teams/access.service";
 
 const delay = (ms: number) => new Promise((res) => setTimeout(res, ms));
@@ -43,7 +43,11 @@ export class RaidersService implements OnModuleInit {
         try {
             this.logger.log("Refreshing raider overviews via cronjob...");
             const overviewsToRefresh = await this.overviewRepository.find({
-                relations: ["raider", "raider.raidTeam"],
+                relations: {
+                    raider: {
+                        raidTeam: true,
+                    },
+                },
                 order: {
                     updatedAt: "ASC",
                 },
@@ -71,7 +75,7 @@ export class RaidersService implements OnModuleInit {
 
     async onModuleInit() {
         this.logger.log("Startup: Discovering the current raid tier through the Blizzard API.");
-        const blizzApi: BlizzardApi = new BlizzardApi(RegionName.us);
+        const blizzApi: BlizzardApi = new BlizzardApi(RegionNameEnum.us);
         const currentRaidTier: RaidTierConfiguration = await blizzApi.getCurrentRaidTier();
         this.logger.log(
             `The current expansion is "${currentRaidTier.expansionName}" (${currentRaidTier.expansionId}).`,
@@ -90,8 +94,10 @@ export class RaidersService implements OnModuleInit {
         );
 
         // Assert character is not already in raid team.
-        const conflictingRaider: Raider = await this.raidersRepository.findOne({
-            raidTeam: raidTeam,
+        const conflictingRaider: Raider = await this.raidersRepository.findOneBy({
+            raidTeam: {
+                id: raidTeam.id,
+            },
             characterName: createRaiderDto.characterName,
             realm: createRaiderDto.realm,
         });
@@ -148,11 +154,10 @@ export class RaidersService implements OnModuleInit {
     async findOne(user: User, raidTeamId: string, raiderId: string): Promise<Raider> {
         await this.accessService.assertUserCanViewRaidTeam(user, raidTeamId);
 
-        return await this.raidersRepository.findOne(raiderId, {
-            where: {
-                raidTeam: {
-                    id: raidTeamId,
-                },
+        return await this.raidersRepository.findOneBy({
+            id: raiderId,
+            raidTeam: {
+                id: raidTeamId,
             },
         });
     }
@@ -163,9 +168,11 @@ export class RaidersService implements OnModuleInit {
             raidTeamId,
         );
 
-        const raider: Raider = await this.raidersRepository.findOne({
+        const raider: Raider = await this.raidersRepository.findOneBy({
             id: raiderId,
-            raidTeam: raidTeam,
+            raidTeam: {
+                id: raidTeam.id,
+            },
         });
         if (!raider) {
             throw new RaiderNotFoundException(
@@ -186,15 +193,15 @@ export class RaidersService implements OnModuleInit {
             raidTeamId,
         );
 
-        const raider: Raider = await this.raidersRepository.findOne(
-            {
+        const raider: Raider = await this.raidersRepository.findOne({
+            where: {
                 id: raiderId,
-                raidTeam: raidTeam,
+                raidTeam: {
+                    id: raidTeam.id,
+                },
             },
-            {
-                relations: ["raidTeam"],
-            },
-        );
+            relations: { raidTeam: true },
+        });
         if (!raider) {
             throw new RaiderNotFoundException(
                 `No raider with id ${raiderId} exists in raid team ${raidTeamId}.`,
@@ -202,8 +209,10 @@ export class RaidersService implements OnModuleInit {
         }
 
         if (useCaching) {
-            const cachedOverview: CachedOverview = await this.overviewRepository.findOne({
-                raider: raider,
+            const cachedOverview: CachedOverview = await this.overviewRepository.findOneBy({
+                raider: {
+                    id: raider.id,
+                },
             });
             const twelveHours = 12 * 60 * 60 * 1000;
             const isOverviewFresh = Date.now() - cachedOverview?.updatedAt?.getTime() < twelveHours;
@@ -257,6 +266,7 @@ export class RaidersService implements OnModuleInit {
         // Include the update timestamp in the returned overview.
         raiderOverview.refreshedAt = updatedAt;
         this.overviewRepository.save({
+            raiderId: raider.id,
             raider: raider,
             cachedOverview: JSON.stringify(raiderOverview),
             // Manually setting updatedAt here so it still refreshes even if the overview did not change.
